Got ScalaTest summary working correctly from Runner.done method, got the master and slave runner wired up correctly.
