+++ scalatest.js/src/main/scala/org/scalatest/tools/SbtReporter.scala
@@ -3,7 +3,7 @@ package org.scalatest.tools
 import sbt.testing.{Status => SbtStatus, _}
 import org.scalatest.Reporter
 
-private class SbtReporter(suiteId: String, fullyQualifiedName: String, fingerprint: Fingerprint, eventHandler: EventHandler, report: Reporter, summaryCounter: SummaryCounter) extends Reporter {
+private class SbtReporter(suiteId: String, fullyQualifiedName: String, fingerprint: Fingerprint, eventHandler: EventHandler, report: Reporter) extends Reporter {
 
   import org.scalatest.events._
 
@@ -32,27 +32,17 @@ private class SbtReporter(suiteId: String, fullyQualifiedName: String, fingerpri
     event match {
       // the results of running an actual test
       case t: TestPending =>
-        summaryCounter.incrementTestsPendingCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getTestSelector(t.suiteId, t.testName), SbtStatus.Pending, new OptionalThrowable, t.duration.getOrElse(0)))
       case t: TestFailed =>
-        summaryCounter.incrementTestsFailedCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getTestSelector(t.suiteId, t.testName), SbtStatus.Failure, getOptionalThrowable(t.throwable), t.duration.getOrElse(0)))
       case t: TestSucceeded =>
-        summaryCounter.incrementTestsSucceededCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getTestSelector(t.suiteId, t.testName), SbtStatus.Success, new OptionalThrowable, t.duration.getOrElse(0)))
       case t: TestIgnored =>
-        summaryCounter.incrementTestsIgnoredCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getTestSelector(t.suiteId, t.testName), SbtStatus.Ignored, new OptionalThrowable, -1))
       case t: TestCanceled =>
-        summaryCounter.incrementTestsCanceledCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getTestSelector(t.suiteId, t.testName), SbtStatus.Canceled, new OptionalThrowable, t.duration.getOrElse(0)))
-      case t: SuiteCompleted =>
-        summaryCounter.incrementSuitesCompletedCount()
       case t: SuiteAborted =>
-        summaryCounter.incrementSuitesAbortedCount()
         eventHandler.handle(ScalaTestSbtEvent(fullyQualifiedName, fingerprint, getSuiteSelector(t.suiteId), SbtStatus.Error, getOptionalThrowable(t.throwable), t.duration.getOrElse(0)))
-      case t: ScopePending =>
-        summaryCounter.incrementScopesPendingCount()
       case _ =>
     }
   }
